header changes the value on file open
