+++ src/App.tsx
@@ -6,31 +6,63 @@ import { CommentCard } from './CommentCard';
 
 import { AppHeader } from './AppHeader';
 
-export default class App extends React.Component {
+import { ISignal } from '@phosphor/signaling';
+
+import { UseSignal } from '@jupyterlab/apputils';
+
+import { ApplicationShell } from '@jupyterlab/application';
+
+import { FocusTracker, Widget } from '@phosphor/widgets';
+
+interface IAppProps {
+  signal?: ISignal<ApplicationShell, FocusTracker.IChangedArgs<Widget>>;
+}
+
+export default class App extends React.Component<IAppProps, {}> {
   constructor(props: any) {
     super(props);
     this.getCommentCards = this.getCommentCards.bind(this);
+  }
+
+  render() {
+    return (
+      <UseSignal signal={this.props.signal}>
+        {(
+          sender: ApplicationShell,
+          args: FocusTracker.IChangedArgs<Widget>
+        ) => {
+          return (
+            <div>
+              {this.checkAppHeader(args)}
+              <AppBody cards={this.getCommentCards(testData)} />
+            </div>
+          );
+        }}
+      </UseSignal>
+    );
   }
 
   getCommentCards(allData: any): React.ReactNode[] {
     let cards: React.ReactNode[] = [];
 
     for (let key in allData) {
-      console.log(key);
       cards.push(<CommentCard data={allData[key]} />);
     }
     return cards;
   }
 
-  render() {
-    return (
-      <div>
-        <div>
-          <AppHeader />
-        </div>
-        <AppBody cards={this.getCommentCards(testData)} />
-      </div>
-    );
+  checkAppHeader(args: any): React.ReactNode {
+    if (
+      args !== undefined &&
+      args !== null &&
+      (args.newValue !== undefined && args.newValue !== null) &&
+      args.newValue.context !== undefined &&
+      args.newValue.context.session !== undefined
+    ) {
+      return <AppHeader header={args.newValue.context.session._name} />;
+    } else {
+      return <AppHeader header={undefined} />;
+    }
   }
 }
