+++ tsconfig.json
@@ -27,6 +27,9 @@
     },
     {
       "path": "../apputils"
+    },
+    {
+      "path": "../docregistry"
     }
   ]
 }
